\name{Homeric-package}
\alias{Homeric-package}
\alias{Homeric}
\docType{package}
\title{Homeric}

\description{Doughnut plots}

\details{
    \tabular{ll}{
    Package: \tab Homeric\cr
    Type: \tab Package\cr
    Version: \tab 0.1-2\cr
    Date: \tab 2016-04-17\cr
    License: \tab GPL-2\cr
    }
}

\author{Maintainer: Lawrence Hudson <l.hudson@nhm.ac.uk>}

\keyword{package}

\examples{
par(mar=rep(0,4), oma=rep(0, 4))
PlotDoughnut(1:5)
}
