\name{Homeric-package}
\alias{Homeric-package}
\alias{Homeric}
\docType{package}
\title{
    \packageTitle{Homeric}
}

\description{
    \packageDescription{Homeric}
}

\details{
    \packageDESCRIPTION{Homeric}
    \packageIndices{Homeric}
}

\author{
    \packageAuthor{Homeric}
    Maintainer: \packageMaintainer{Homeric}
}

\keyword{package}

\examples{
par(mar=rep(0,4), oma=rep(0, 4))
PlotDoughnut(1:5)
}
